server/requirements.txt
Django==2.0.13
djangorestframework==3.9.2
pytz==2018.9
